py/requirements.txt
mysql
mysql.connector
python-decouple
